% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_to_long .R
\name{board_to_long}
\alias{board_to_long}
\title{Transform board to long dataframe}
\usage{
board_to_long(board)
}
\arguments{
\item{board}{An \eqn{ (\text{size}+2) \times (\text{size}+2) } matrix.}
}
\value{
An dataframe for plotting board.
}
\description{
Transform board to long dataframe
}
\examples{
board_to_long(board)
}
