% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hex_winner.R
\name{check_hex_winner}
\alias{check_hex_winner}
\title{Check hex winner}
\usage{
check_hex_winner(board)
}
\arguments{
\item{board}{An \eqn{ (\text{size}+2) \times (\text{size}+2) } matrix.}
}
\value{
An integer of winner number.
}
\description{
Check hex winner
}
\examples{
check_hex_winner(board)
}
