% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_presets.R
\name{colour_presets}
\alias{colour_presets}
\title{Helper function to show RGB colour presets at 0-1 scale}
\usage{
colour_presets()
}
\value{
A data frame
}
\description{
Helper function to show RGB colour presets at 0-1 scale
}
\examples{
colour_presets()

}
