% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs.R
\name{dfs}
\alias{dfs}
\title{dfs}
\usage{
dfs(board, player, visited, row, col)
}
\arguments{
\item{board}{An \eqn{ (\text{size}+2) \times (\text{size}+2) } matrix.}

\item{player}{An integer of player number.}

\item{visited}{An \eqn{ (\text{size}+2) \times (\text{size}+2) } matrix.}

\item{row}{An integer. This parameter shows row number of board.}

\item{col}{An integer. This parameter shows col number of board.}
}
\value{
A Logical.
}
\description{
dfs
}
\examples{
dfs(board, 1, visited, 2, 3)
}
