% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighbors.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get neighbors of hex}
\usage{
get_neighbors(row, col, board_size)
}
\arguments{
\item{row}{An integer. This parameter shows row number of board.}

\item{col}{An integer. This parameter shows col number of board.}

\item{board_size}{An integer. This parameter shows board size.}
}
\value{
A list of position of board.
}
\description{
Get neighbors of hex
}
\examples{
neighbors <- get_neighbors(1, 1, 5)
}
