% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_block.R
\name{hex_block}
\alias{hex_block}
\title{Display a hex colour as a plot}
\usage{
hex_block(red_val, green_val, blue_val)
}
\arguments{
\item{red_val}{a value between 0-1 for the red spectrum}

\item{green_val}{a value between 0-1 for the green spectrum}

\item{blue_val}{a value between 0-1 for the blue spectrum}
}
\value{
a hex code (character) and a plot (ggplot2 object)
}
\description{
Display a hex colour as a plot
}
\examples{
hex_block(0.9, 0.1, 0.5)
}
