% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_code.R
\name{hex_code}
\alias{hex_code}
\title{Generate a hex code}
\usage{
hex_code(red_val, green_val, blue_val)
}
\arguments{
\item{red_val}{a value between 0-1 for the red spectrum}

\item{green_val}{a value between 0-1 for the green spectrum}

\item{blue_val}{a value between 0-1 for the blue spectrum}
}
\value{
a hex code character vector
}
\description{
Generate a hex code
}
\examples{
hex_code(0.9, 0.1, 0.5)
}
