% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_digit.R
\name{hex_digit}
\alias{hex_digit}
\title{Convert an RGB digit into a hex digits (helper function)}
\usage{
hex_digit(rgb_digit)
}
\arguments{
\item{rgb_digit}{a raw RGB digit value between 0 and 255}
}
\value{
a double
}
\description{
Convert an RGB digit into a hex digits (helper function)
}
\examples{
hex_digit(123)
}
