% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_translation.R
\name{hex_translation}
\alias{hex_translation}
\title{Show RGB digits to hex digits translatation}
\usage{
hex_translation()
}
\value{
a named vector
}
\description{
Show RGB digits to hex digits translatation
}
\examples{
hex_translation()
}
