% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_hex_board.R
\name{initialize_hex_board}
\alias{initialize_hex_board}
\title{Initialize a board}
\usage{
initialize_hex_board(size = 11)
}
\arguments{
\item{size}{An integer. This parameter shows board size.}
}
\value{
An \eqn{ (\text{size}+2) \times (\text{size}+2) } matrix.
}
\description{
Initialize a board
}
\examples{
board <- initialize_hex_board(5)
}
