% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hex_board.R
\name{plot_hex_board}
\alias{plot_hex_board}
\title{Plot hex board}
\usage{
plot_hex_board(board)
}
\arguments{
\item{board}{An \eqn{ (\text{size}+2) \times (\text{size}+2) } matrix.}
}
\value{
A plot.
}
\description{
Plot hex board
}
\examples{
plot_hex_board(board)
}
