% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_hexr.R
\name{shiny_hexr}
\alias{shiny_hexr}
\title{A simple colour picker shiny app}
\usage{
shiny_hexr(...)
}
\arguments{
\item{...}{Additional shiny arguments}
}
\value{
A shiny runtime
}
\description{
A simple colour picker shiny app
}
\examples{
 \dontrun{
 shiny_hexr()
}
}
